+++ rs-week1/src/App.tsx
@@ -6,7 +6,6 @@ import About from './components/pages/About';
 import Page404 from './components/pages/Page404';
 
 interface IAppState {
-  searchTerm: string;
   namePage: string;
 }
 
@@ -14,33 +13,21 @@ class App extends Component<Record<string, never>, IAppState> {
   constructor(props: Record<string, never>) {
     super(props);
     this.state = {
-      searchTerm: '',
       namePage: '111',
     };
   }
-
-  handleSearch = (searchTerm: string) => {
-    this.setState({ searchTerm });
-  };
 
   handleNamePage = (namePage: string) => {
     this.setState({ namePage });
   };
 
   render() {
-    const { searchTerm, namePage } = this.state;
+    const { namePage } = this.state;
     return (
       <>
-        <Header onSearch={this.handleSearch} namePageTerm={namePage} />
+        <Header namePageTerm={namePage} />
         <Routes>
-          <Route
-            path="/"
-            element={<Home searchTerm={searchTerm} onChangeNamePage={this.handleNamePage} />}
-          />
-          <Route
-            path="/home"
-            element={<Home searchTerm={searchTerm} onChangeNamePage={this.handleNamePage} />}
-          />
+          <Route path="/" element={<Home onChangeNamePage={this.handleNamePage} />} />
           <Route path="/about" element={<About onChangeNamePage={this.handleNamePage} />} />
           <Route path="*" element={<Page404 onChangeNamePage={this.handleNamePage} />} />
         </Routes>
