+++ rs-week1/src/components/pages/Home.css
@@ -7,6 +7,54 @@
   padding: 0 1rem;
 }
 
+.card-search {
+  margin-bottom: 0.5rem;
+  background-color: #343a40;
+  padding: 0.5rem;
+}
+
+.card-search form {
+  display: flex;
+  align-items: center;
+}
+
+.card-search input[type="text"] {
+  width: 200px;
+  height: 20px;
+  margin-right: 10px;
+  border: none;
+  border-radius: 20px;
+  padding: 5px 10px;
+  font-size: 16px;
+  color: #333;
+  background-color: #fff;
+}
+
+.card-search button[type="submit"] {
+  width: 80px;
+  height: 30px;
+  border: none;
+  border-radius: 20px;
+  font-size: 16px;
+  font-weight: bold;
+  color: #fff;
+  background-color: #007bff;
+  cursor: pointer;
+  transition: all 0.1s ease-in-out;
+}
+
+.card-search button[type="submit"]:hover {
+  background-color: #0069d9;
+}
+
+.card-search button[type="submit"]:active {
+  background-color: rgb(101, 236, 83);
+}
+
+.card-search {
+  flex-grow: 1;
+}
+
 
 @media screen and (max-width: 1024px) {
   .product-cards-container {
