+++ rs-week1/src/components/pages/Home.tsx
@@ -1,4 +1,4 @@
-// import products from '../dataProducts/products';
+import products from '../dataProducts/products';
 import ProductCard from '../ProductCard';
 import React, { Component } from 'react';
 import './Home.css';
@@ -6,65 +6,83 @@ import { IProduct } from 'components/types/interface';
 
 interface IState {
   products: IProduct[];
-  searchTerm: string;
-  isLoading: boolean;
+  searchInput: string;
 }
 
 interface IHomePageProps {
-  searchTerm: string;
   onChangeNamePage: (namePage: string) => void;
 }
 
 class Home extends Component<IHomePageProps, IState> {
+  private beforeUnloadListener: ((event: BeforeUnloadEvent) => void) | null = null;
   constructor(props: IHomePageProps) {
     super(props);
     this.state = {
-      products: [],
-      searchTerm: props.searchTerm,
-      isLoading: true,
+      products: products,
+      searchInput: '',
     };
   }
 
-  async componentDidMount() {
+  componentDidMount() {
     this.props.onChangeNamePage('Home Page');
-    try {
-      const response = await fetch('https://dummyjson.com/products?limit=12');
-      const resProducts = await response.json();
-      const products = resProducts.products;
-      this.setState({ products, isLoading: false });
-    } catch (error) {
-      console.error('Error fetching products:', error);
-      this.setState({ isLoading: false });
+    const searchTerm = localStorage.getItem('searchTerm');
+    if (searchTerm) {
+      this.setState({ searchInput: searchTerm });
+    }
+    // Добавляем обработчик события beforeunload
+    this.beforeUnloadListener = () => {
+      localStorage.setItem('searchTerm', this.state.searchInput);
+    };
+    window.addEventListener('beforeunload', this.beforeUnloadListener);
+  }
+
+  componentWillUnmount() {
+    localStorage.setItem('searchTerm', this.state.searchInput);
+    if (this.beforeUnloadListener) {
+      window.removeEventListener('beforeunload', this.beforeUnloadListener);
     }
   }
 
-  componentDidUpdate(prevProps: IHomePageProps) {
-    if (prevProps.searchTerm !== this.props.searchTerm) {
-      this.setState({ searchTerm: this.props.searchTerm });
-    }
-  }
+  componentDidUpdate() {}
+
+  handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const { value } = event.target;
+    this.setState({ searchInput: value });
+  };
+
+  handleFormSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+  };
 
   render() {
-    const { products, searchTerm, isLoading } = this.state;
+    const { products } = this.state;
+    const { searchInput } = this.state;
 
     const filteredProducts = products.filter((product) => {
-      return product.title.toLowerCase().includes(searchTerm.toLowerCase());
+      return product.title.toLowerCase().includes(searchInput.toLowerCase());
     });
 
     return (
       <>
+        <div className="card-search">
+          <form onSubmit={this.handleFormSubmit}>
+            <input
+              type="text"
+              placeholder="Search..."
+              value={searchInput}
+              onChange={this.handleInputChange}
+            />
+            <button type="submit">Search</button>
+          </form>
+        </div>
         <h3>Home page</h3>
-        {isLoading ? (
-          <p>Loading...</p>
-        ) : (
-          <div className="product-cards-container">
-            {filteredProducts.length ? (
-              filteredProducts.map((product) => <ProductCard product={product} key={product.id} />)
-            ) : (
-              <p>No found product ...</p>
-            )}
-          </div>
-        )}
+        <div className="product-cards-container">
+          {filteredProducts.length ? (
+            filteredProducts.map((product) => <ProductCard product={product} key={product.id} />)
+          ) : (
+            <p>No found product ...</p>
+          )}
+        </div>
       </>
     );
   }
